+++ HISTORY.md
@@ -8,9 +8,12 @@ All notable changes to this project will be documented in this file.
 
 * BREAKING Changes
   * Dropped support for `node<20.18.0` ([#1362] via [#1365])
+* Dependencies
+  * Upgraded runtime-dependency `@cyclonedx/cyclonedx-library@^8.0.0`, was `@^7.0.0` (via [#1367])
 
 [#1362]: https://github.com/CycloneDX/cyclonedx-webpack-plugin/issues/1362
 [#1365]: https://github.com/CycloneDX/cyclonedx-webpack-plugin/pull/1365
+[#1367]: https://github.com/CycloneDX/cyclonedx-webpack-plugin/pull/1367
 
 ## 4.0.1 - 2025-01-29